+++ src/models/deviceSettings/shutterSettings.ts
@@ -10,6 +10,11 @@ export class ShutterSettings extends DeviceSettings {
    */
   public direction?: number;
   /**
+   * The desired position, when the shutter should perform a heat reduction.
+   * @type {number}
+   */
+  public heatReductionPosition: number = 40;
+  /**
    * Some shutter give no position feedback on their own, so by knowing the durations in either direction,
    * we can programmatically trigger the callbacks.
    * @type {boolean}