+++ src/server/devices/groups/windowGroup.ts
@@ -1,4 +1,4 @@
-import { LogLevel, RoomBase, TimeCallback, TimeCallbackType, TimeOfDay } from '../../../models';
+import { LogLevel, RoomBase, ShutterSettings, TimeCallback, TimeCallbackType, TimeOfDay } from '../../../models';
 import { ShutterService, TimeCallbackService, Utils, WeatherService } from '../../services';
 import { Window } from './Window';
 import { WindowPosition } from '../models';
@@ -6,7 +6,6 @@ import { BaseGroup } from './base-group';
 import { GroupType } from './group-type';
 import { DeviceList } from '../device-list';
 import { DeviceClusterType } from '../device-cluster-type';
-import { iShutter } from '../baseDeviceInterfaces';
 
 export class WindowGroup extends BaseGroup {
   public sunriseShutterCallback: TimeCallback | undefined;
@@ -79,8 +78,8 @@ export class WindowGroup extends BaseGroup {
     const timeOfDay: TimeOfDay = TimeCallbackService.dayType(room.settings.rolloOffset);
     const darkOutside: boolean = TimeCallbackService.darkOutsideOrNight(timeOfDay);
     this.windows.forEach((f) => {
-      const shutter: iShutter[] = f.getShutter();
-      if (shutter.length === 0) {
+      const shutterSettings: ShutterSettings | undefined = f.getShutter()?.[0]?.settings;
+      if (!shutterSettings) {
         return;
       }
       if (darkOutside) {
@@ -94,7 +93,7 @@ export class WindowGroup extends BaseGroup {
           room.HeatGroup?.desiredTemp ?? -99,
           room.HeatGroup?.temperature ?? -99,
           this.log.bind(this),
-          shutter[0]?.settings.direction,
+          shutterSettings,
         );
       }
       if (f.griffeInPosition(WindowPosition.offen) > 0 && desiredPos < 100) {